+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to the "package-json-upgrade" extension will be documented i
 
 ## Unreleased changes
 
+- Fix so "~5.0.0" is not displayed as "version not found" if some version matching the range exists, such as "5.0.1"
 - Fix so "workspace:" syntax does not display "Dependency not found" message
+- Fix certain non-registry versions erronously being fetched
 
 ## 3.6.0